requirements.txt
python-dotenv
streamlit
langchain-groq
langchain-community
langchain-core
langchain-text-splitters
langchain
langchain-huggingface
faiss-cpu
PyPDF2
sentence-transformers
